requirements.txt
python-telegram-bot==1.2.1
python-dotenv==22.5
